Double -> Long

+++ src/main/java/pt/ulisboa/tecnico/meic/cnv/Metric.java
@@ -5,11 +5,11 @@ public class Metric {
     private static final double TAKEN_RATIO = 1d / 3d;
     private static final double NOT_TAKEN_RATIO = 1d / 3d;
 
-    private double mCount;
-    private double taken;
-    private double notTaken;
+    private long mCount;
+    private long taken;
+    private long notTaken;
 
-    public Metric(double mCount, double taken, double notTaken) {
+    public Metric(long mCount, long taken, long notTaken) {
         this.mCount = mCount;
         this.taken = taken;
         this.notTaken = notTaken;
